Revert extra change and create separte tests for auto-renew e2e
