+++ package.json
@@ -48,15 +48,13 @@
   "devDependencies": {
     "@playwright/test": "^1.45.0",
     "@types/node": "^20.11.17",
-    "@types/sinon": "^17.0.3",
     "browserstack-local": "^1.4.8",
     "dotenv": "^16.0.3",
     "eslint": "^8.44.0",
     "eslint-config-prettier": "^8.8.0",
     "husky": "^8.0.3",
     "playwright": "^1.45.0",
-    "prettier": "^3.0.0",
-    "sinon": "^18.0.0"
+    "prettier": "^3.0.0"
   },
   "eslintConfig": {
     "parser": "@typescript-eslint/parser",
